Add files via upload
It maintains information on drivers including their driving licence number, name, number of points on the licence and contact number so they can contact them in future for offers. It also maintains information on all cars in its system. Details stored about the car are registration number, year of manufacture, engine size and whether it has passed its NCT test.

The Insurance company can create policies for drivers registered with the company if their car is also in the system (policies apply to one driver and one car only – a policy has a car and a driver). Policies can be created and stored on the system if the driver has under 12 points on their licence and if the car has passed its NCT. It also stores a policy id, premium and number of claims for each policy. 

The application should allow the manager of the insurance to: 
•	Create & add new drivers, cars and policies
•	Delete a policy
•	Increase premiums annually by 5%. 
•	Show the total number of claims for all policies in the list.
•	Show the total amount made by the Company from all policy premiums.